Let sixty outlets outrank one model's guess
A phone launch covered by 65 sources rendered as the smallest card. The size
gate read `importance >= 4 && source_count >= 3`, an AND, so importance was a
hard veto — and the model calls a scheduled product launch routine, because
against a front page it is. Breadth now promotes a card on its own.

Two matching corrections to the ranking it sits beside: corroboration is
clamped, having silently passed 1.0 beyond 25 sources and carried the whole
product above its documented range, and the importance term is mapped to
0.6-1.0 instead of 0.2-1.0, which was a 5x swing that settled the order before
recency or corroboration got a say.

Also: the header's month is abbreviated, its place reads as city and
two-letter country rather than the full administrative label, and the reels
foot keeps one time value instead of two.

+++ src/components/HeaderAside.tsx
@@ -19,7 +19,7 @@ export function HeaderAside() {
   const [here, setHere] = useState('');
 
   useEffect(() => {
-    setToday(new Date().toLocaleDateString('en-GB', { weekday: 'short', day: 'numeric', month: 'long' }));
+    setToday(new Date().toLocaleDateString('en-GB', { weekday: 'short', day: 'numeric', month: 'short' }));
     try { setHere(localStorage.getItem(KEY) ?? ''); } catch { /* first visit */ }
 
     let live = true;
